Starts work on refactoring parsing exceptions
